rename to `powerfit_rs`
