+++ src/powerfit_em/__init__.py
@@ -1,5 +1,5 @@
 from .helpers import determine_core_indices
-from .powerfitrs import __version__
+from .powerfit_rs import __version__
 from .rotations import proportional_orientations, quat_to_rotmat
 from .structure import Structure
 from .volume import Volume, structure_to_shape_like
